backend/requirements.txt
Flask
Flask-Cors
Flask-MySQLdb
bcrypt
PyJWT
python-dotenv
